src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rank-bm25==0.2.2
PyMuPDF==1.24.3
langdetect==1.0.9
deep-translator==1.11.4
reportlab==4.1.0
numpy==1.26.4
scikit-learn==1.4.2
httpx==0.27.0
python-dotenv==1.0.1
google-cloud-aiplatform>=1.49.0
google-cloud-texttospeech
